Added rasterstats functions to seperate file
Added rasterstats functions to seperate file
